Update dependency name

+++ Sources/FileLogging/Handlers/SizeRotatingFileLogHandler.swift
@@ -1,6 +1,6 @@
 import Foundation
 import Logging
-import TrailBlazer
+import Pathman
 
 public struct SizeRotatingFileLogHandler: RotatingFileLogHandler {
     private let filename: String
@@ -41,7 +41,7 @@ public struct SizeRotatingFileLogHandler: RotatingFileLogHandler {
 
     public func rotate(message: Data) -> String? {
         let msgSize = message.count
-        guard logFile.size + msgSize > maxSize else { return nil }
+        guard logFile.size + Int64(msgSize) > maxSize else { return nil }
 
         guard msgSize <= maxSize else {
             fatalError("Message is larger than maximum byte size allowed per file rotation (\(msgSize) > \(maxSize))")
